Struggling With Modify Comment

+++ demo/src/main/java/edu/test/demo/controller/TestController2.java
@@ -1,16 +1,19 @@
 package edu.test.demo.controller;
 
 import java.io.*;
+import java.lang.ProcessBuilder.Redirect;
 import java.util.*;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.web.WebProperties.Resources.Chain.Strategy.Content;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.util.JavaScriptUtils;
 
 import edu.test.demo.service.CocommentService;
 import edu.test.demo.service.CommentService;
@@ -49,15 +52,23 @@ public class TestController2 {
 
 //modify user info
 	@GetMapping("/userinfo/modify")
-	public String modifyUserInfoPage(@RequestParam Integer user_id) {
-		return "main/userModify";
+	public String modifyUserInfoPage(Model model,HttpSession session, @RequestParam Integer user_id) {
+		//if (login user != targeted user)can't modify information
+		int session_user_id = ((UserVO)session.getAttribute("user")).getUser_id();
+		if(session_user_id == user_id)
+			return "main/userModify";
+		else {
+			model.addAttribute("msg","Unauthorized User!");
+			return "main/fail";
+		}
 	}
 	@PostMapping("/userinfo/modify")
-	public String modifyUserInfo(UserVO userVO, UserCharacterVO userCharacterVO, HttpSession session) {
+	public String modifyUserInfo(Model model, UserVO userVO, UserCharacterVO userCharacterVO, HttpSession session, HttpServletRequest request, @RequestParam(value="file") MultipartFile file) {
 		try {
-			userService.modifyUser(userVO,userCharacterVO,session);
+			userService.modifyUser(userVO,userCharacterVO,session, request, file);
 			return "main/success";
 		}catch (Exception e) {
+			model.addAttribute("msg",e);
 			System.out.println(e);
 			return "main/fail";
 		}
@@ -65,26 +76,59 @@ public class TestController2 {
 	
 //댓글입력
 	@PostMapping("/comment")
-	public String PostComment(CommentVO commentVO, HttpSession session){
+	public String PostComment(Model model, CommentVO commentVO, HttpSession session){
 		try {
 			commentService.insertComment(commentVO);
 			return "main/success";
 		}catch (Exception e) {
+			model.addAttribute("msg",e);
 			System.out.println(e);
 			return "main/fail";
 		}
 	}
-//댓글 삭제(해당 댓글의 status를 2로 바꿈)
+	//댓글 삭제(해당 댓글의 status를 2로 바꿈)
 	@PostMapping("/delco")
-	public String deleteComment(int comment_id) {
+	public String deleteComment(Model model, HttpSession session, int comment_id) {
 		try {
-			commentService.deleteComment(comment_id);
-			return "main/success";
+			//if (login user != commented user) can't delete.
+			int session_user_id = ((UserVO)session.getAttribute("user")).getUser_id();
+			int user_id = commentService.selectCommentByCommentId(comment_id).getComment_id_from();
+			if(session_user_id==user_id) {
+				commentService.deleteComment(comment_id);
+				return "main/success";
+			}else {
+				System.out.println("YOU. ARE. NOT. THAT. USER!");
+				model.addAttribute("msg","Unauthorized User");
+				return "main/fail";
+			}
 		} catch (Exception e) {
+			model.addAttribute("msg",e);
 			System.out.println(e);
 			return "main/fail";
 		}
 	}
+	//댓글 수정(해당 댓글의 status를 1로 바꾸고 log를 새로 생성)
+	@PostMapping("/modico")
+	public String modifyComment(Model model,HttpSession session, int comment_id, String comment_contents) {
+		try {
+			//if (login user != commented user) can't modify.
+			int session_user_id = ((UserVO)session.getAttribute("user")).getUser_id();					//세션 유저 id
+			int user_id = commentService.selectCommentByCommentId(comment_id).getComment_id_from();		//댓글을 쓴 유저 id
+			if(session_user_id==user_id) {
+				commentService.modifyComment(comment_id, comment_contents);
+				return "main/success";
+			}else {
+				System.out.println("YOU. ARE. NOT. THAT. USER!");
+				model.addAttribute("msg","Unauthorized User");
+				return "main/fail";
+			}
+		} catch (Exception e) {
+			model.addAttribute("msg",e);
+			System.out.println(e);
+			return "main/fail";
+		}
+	}
+	
 	
 //대댓글입력
 	@PostMapping("/cocomment")
@@ -94,13 +138,44 @@ public class TestController2 {
 		cocommentService.insertCocomment(cocommentVO);
 		return "main/success";
 	}
-//대댓글 삭제(해당 대댓글의 status를 2로 바꿈)
+	//대댓글 삭제(해당 대댓글의 status를 2로 바꿈)
 	@PostMapping("/delcoco")
-	public String deleteCocomment(int cocomment_id) {
+	public String deleteCocomment(Model model, HttpSession session, int cocomment_id) {
 		try {
-			cocommentService.deleteCocomment(cocomment_id);
-			return "main/success";
+			//if (login user != cocommented user) can't delete.
+			int session_user_id = ((UserVO)session.getAttribute("user")).getUser_id();
+			int user_id = cocommentService.selectCocommentByCocommentId(cocomment_id).getCocomment_id_from();
+			if(session_user_id==user_id) {
+				cocommentService.deleteCocomment(cocomment_id);
+				return "main/success";
+			}else {
+				System.out.println("YOU. ARE. NOT. THAT. USER!");
+				model.addAttribute("msg","Unauthorized User");
+				return "main/fail";
+			}
 		} catch (Exception e) {
+			model.addAttribute("msg",e);
+			System.out.println(e);
+			return "main/fail";
+		}
+	}
+	//대댓글 수정
+	@PostMapping("/modicoco")
+	public String modifyCocomment(Model model,HttpSession session, int cocomment_id, String cocomment_contents) {
+		try {
+			// if (login user != cocommented user) can't modify.
+			int session_user_id = ((UserVO)session.getAttribute("user")).getUser_id();					//세션 유저 id
+			int user_id = cocommentService.selectCocommentByCocommentId(cocomment_id).getCocomment_id_from();		//대댓글을 쓴 유저 id
+			if(session_user_id==user_id) {
+				cocommentService.modifyComment(cocomment_id, cocomment_contents);
+				return "main/success";
+			}else {
+				System.out.println("YOU. ARE. NOT. THAT. USER!");
+				model.addAttribute("msg","Unauthorized User");
+				return "main/fail";
+			}
+		} catch (Exception e) {
+			model.addAttribute("msg",e);
 			System.out.println(e);
 			return "main/fail";
 		}
@@ -126,13 +201,14 @@ public class TestController2 {
 	}
 	
 	@PostMapping("/login")
-	public String loginPost(String user_email, String user_pw, HttpSession session) {
+	public String loginPost(Model model, String user_email, String user_pw, HttpSession session) {
 		UserVO user = userService.loginCheck(user_email, user_pw); 
-		System.out.println(user_email +" / "+ user_pw);
 		if (user == null) {
+			model.addAttribute("msg","Log In First");
 			return "main/fail";
 		}else {
 			session.setAttribute("user", user);
+			session.setAttribute("userCharacter", userCharacterService.selectUserCharacterByUserId(user.getUser_id()));
 			return "main/success";
 		}
